+++ server/src/main/java/org/elasticsearch/index/fielddata/plain/SortedSetOrdinalsIndexFieldData.java
@@ -19,27 +19,20 @@
 
 package org.elasticsearch.index.fielddata.plain;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.OrdinalMap;
 import org.apache.lucene.index.SortedSetDocValues;
 import org.apache.lucene.search.SortField;
 import org.apache.lucene.search.SortedSetSelector;
 import org.apache.lucene.search.SortedSetSortField;
-import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.util.BigArrays;
 import org.elasticsearch.index.fielddata.IndexFieldData;
 import org.elasticsearch.index.fielddata.IndexFieldData.XFieldComparatorSource.Nested;
 import org.elasticsearch.index.fielddata.IndexFieldDataCache;
-import org.elasticsearch.index.fielddata.IndexOrdinalsFieldData;
 import org.elasticsearch.index.fielddata.LeafOrdinalsFieldData;
 import org.elasticsearch.index.fielddata.ScriptDocValues;
 import org.elasticsearch.index.fielddata.fieldcomparator.BytesRefFieldComparatorSource;
-import org.elasticsearch.index.fielddata.ordinals.GlobalOrdinalsBuilder;
-import org.elasticsearch.index.fielddata.ordinals.GlobalOrdinalsIndexFieldData;
 import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.indices.breaker.CircuitBreakerService;
 import org.elasticsearch.search.DocValueFormat;
@@ -48,10 +41,9 @@ import org.elasticsearch.search.aggregations.support.ValuesSourceType;
 import org.elasticsearch.search.sort.BucketedSort;
 import org.elasticsearch.search.sort.SortOrder;
 
-import java.io.IOException;
 import java.util.function.Function;
 
-public class SortedSetOrdinalsIndexFieldData implements IndexOrdinalsFieldData {
+public class SortedSetOrdinalsIndexFieldData extends AbstractIndexOrdinalsFieldData {
 
     public static class Builder implements IndexFieldData.Builder {
         private final String name;
@@ -78,13 +70,6 @@ public class SortedSetOrdinalsIndexFieldData implements IndexOrdinalsFieldData {
         }
     }
 
-    protected final String fieldName;
-    private final IndexFieldDataCache cache;
-    private final CircuitBreakerService breakerService;
-    private final Function<SortedSetDocValues, ScriptDocValues<?>> scriptFunction;
-    private final ValuesSourceType valuesSourceType;
-    private static final Logger logger = LogManager.getLogger(SortedSetOrdinalsIndexFieldData.class);
-
     public SortedSetOrdinalsIndexFieldData(
         IndexFieldDataCache cache,
         String fieldName,
@@ -92,21 +77,7 @@ public class SortedSetOrdinalsIndexFieldData implements IndexOrdinalsFieldData {
         CircuitBreakerService breakerService,
         Function<SortedSetDocValues, ScriptDocValues<?>> scriptFunction
     ) {
-        this.fieldName = fieldName;
-        this.valuesSourceType = valuesSourceType;
-        this.cache = cache;
-        this.breakerService = breakerService;
-        this.scriptFunction = scriptFunction;
-    }
-
-    @Override
-    public final String getFieldName() {
-        return fieldName;
-    }
-
-    @Override
-    public ValuesSourceType getValuesSourceType() {
-        return valuesSourceType;
+        super(fieldName, valuesSourceType, cache, breakerService, scriptFunction);
     }
 
     @Override
@@ -121,7 +92,7 @@ public class SortedSetOrdinalsIndexFieldData implements IndexOrdinalsFieldData {
                 (source.sortMissingLast(missingValue) == false && source.sortMissingFirst(missingValue) == false)) {
             return new SortField(getFieldName(), source, reverse);
         }
-        SortField sortField = new SortedSetSortField(fieldName, reverse,
+        SortField sortField = new SortedSetSortField(getFieldName(), reverse,
             sortMode == MultiValueMode.MAX ? SortedSetSelector.Type.MAX : SortedSetSelector.Type.MIN);
         sortField.setMissingValue(source.sortMissingLast(missingValue) ^ reverse ?
             SortedSetSortField.STRING_LAST : SortedSetSortField.STRING_FIRST);
@@ -136,63 +107,12 @@ public class SortedSetOrdinalsIndexFieldData implements IndexOrdinalsFieldData {
 
     @Override
     public LeafOrdinalsFieldData load(LeafReaderContext context) {
-        return new SortedSetBytesLeafFieldData(context.reader(), fieldName, scriptFunction);
+        return new SortedSetBytesLeafFieldData(context.reader(), getFieldName(), scriptFunction);
     }
 
     @Override
-    public LeafOrdinalsFieldData loadDirect(LeafReaderContext context) throws Exception {
+    public LeafOrdinalsFieldData loadDirect(LeafReaderContext context) {
         return load(context);
-    }
-
-    @Override
-    public IndexOrdinalsFieldData loadGlobal(DirectoryReader indexReader) {
-        IndexOrdinalsFieldData fieldData = loadGlobalInternal(indexReader);
-        if (fieldData instanceof GlobalOrdinalsIndexFieldData) {
-            // we create a new instance of the cached value for each consumer in order
-            // to avoid creating new TermsEnums for each segment in the cached instance
-            return ((GlobalOrdinalsIndexFieldData) fieldData).newConsumer(indexReader);
-        } else {
-            return fieldData;
-        }
-    }
-
-    private IndexOrdinalsFieldData loadGlobalInternal(DirectoryReader indexReader) {
-        if (indexReader.leaves().size() <= 1) {
-            // ordinals are already global
-            return this;
-        }
-        boolean fieldFound = false;
-        for (LeafReaderContext context : indexReader.leaves()) {
-            if (context.reader().getFieldInfos().fieldInfo(getFieldName()) != null) {
-                fieldFound = true;
-                break;
-            }
-        }
-        if (fieldFound == false) {
-            // Some directory readers may be wrapped and report different set of fields and use the same cache key.
-            // If a field can't be found then it doesn't mean it isn't there,
-            // so if a field doesn't exist then we don't cache it and just return an empty field data instance.
-            // The next time the field is found, we do cache.
-            try {
-                return GlobalOrdinalsBuilder.buildEmpty(indexReader, this);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
-        try {
-            return cache.load(indexReader, this);
-        } catch (Exception e) {
-            if (e instanceof ElasticsearchException) {
-                throw (ElasticsearchException) e;
-            } else {
-                throw new ElasticsearchException(e);
-            }
-        }
-    }
-
-    @Override
-    public IndexOrdinalsFieldData loadGlobalDirect(DirectoryReader indexReader) throws Exception {
-        return GlobalOrdinalsBuilder.build(indexReader, this, breakerService, logger, scriptFunction);
     }
 
     @Override
